+++ misc/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## unreleased
 
+-   fix live page fullscreen button
+
 ## `v0.3.5`
 
 -   move settings documentation into the settings page